% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_tbl.R
\name{crosstab}
\alias{crosstab}
\title{creating cross tabulation}
\usage{
crosstab(
  ...,
  dec.places = NULL,
  type = NULL,
  style = "wide",
  row.vars = NULL,
  col.vars = NULL,
  percentages = TRUE,
  addmargins = TRUE,
  subtotals = TRUE
)
}
\arguments{
\item{...}{The dot dot dot operation.}

\item{dec.places}{Number of decimal places.}

\item{type}{Type of Cross Tabulation.}

\item{style}{Style of output.}

\item{row.vars}{Variable input for row-wise operations.}

\item{col.vars}{Variable input for column-wise operations.}

\item{percentages}{Option to display percentages.}

\item{addmargins}{adding margins.}

\item{subtotals}{Option to display subtotal to each group.}
}
\description{
creating cross tabulation
}
